requirements.txt
python-miio
paho-mqtt